Updated some configuration

+++ cucumber-tests-platform/src/test/java/com/alliander/osgp/platform/cucumber/steps/database/DatabaseSteps.java
@@ -54,6 +54,8 @@ public class DatabaseSteps {
         cleanRepoSerializable(this.manufacturerRepo);
 		for(Organisation org : this.organizationRepo.findAll()) {
 			if (!org.getOrganisationIdentification().equals("test-org") && 
+					!org.getOrganisationIdentification().equals("FlexOvlProject") &&
+					!org.getOrganisationIdentification().equals("GemeenteArnhem") &&
 					!org.getOrganisationIdentification().equals("LianderNetManagement")) {
 				this.organizationRepo.delete(org);
 			}
@@ -104,14 +106,20 @@ public class DatabaseSteps {
 	}
 
 	public void prepareDatabaseForScenario() {
+		Logger LOGGER = LoggerFactory.getLogger(this.getClass());
+		LOGGER.info(this.getClass() + ": prepareDatabaseForScenario");
+		
         cleanRepoAbstractEntity(this.oslpDeviceRepo);
         cleanRepoAbstractEntity(this.deviceAuthorizationRepo);
         cleanRepoSerializable(this.deviceRepo);
         cleanRepoSerializable(this.deviceModelRepo);
         cleanRepoSerializable(this.manufacturerRepo);
 		for(Organisation org : this.organizationRepo.findAll()) {
-			if (!org.getOrganisationIdentification().equals("test-org") && 
-					!org.getOrganisationIdentification().equals("LianderNetManagement")) {
+			String orgName = org.getOrganisationIdentification(); 
+			if (!orgName.equals("test-org") && 
+					!orgName.equals("FlexOvlProject") &&
+					!orgName.equals("GemeenteArnhem") &&
+					!orgName.equals("LianderNetManagement")) {
 				this.organizationRepo.delete(org);
 			}
 		}
